requirements.txt
numpy
pandas
yfinance
pyyaml
scikit-learn
matplotlib
